refactor: convert server components to client-side auth for static export
Dashboard layout and login page used server-side `await auth()` with
`redirect()`, which blocks Next.js static export. Converted both to
client components using `useSession()` + `useRouter()` for auth checks.

Also split the game detail page into a server wrapper (page.tsx with
`generateStaticParams`) and client component (game-detail-page.tsx) so
the dynamic [appId] route is compatible with static export.

Added conditional `output: 'export'`, `basePath`, and `images.unoptimized`
to next.config.ts, gated behind `STATIC_EXPORT=true` env var so the
regular build and CI are unaffected.
