+++ next.config.ts
@@ -1,7 +1,14 @@
 import type { NextConfig } from "next";
 
+const isStaticExport = process.env.STATIC_EXPORT === "true";
+
 const nextConfig: NextConfig = {
+  ...(isStaticExport && {
+    output: "export",
+    basePath: process.env.NEXT_PUBLIC_BASE_PATH || "",
+  }),
   images: {
+    unoptimized: isStaticExport,
     remotePatterns: [
       {
         protocol: "https",
